Endrer tilbake til en scrape
